requirements.txt
# MCP Server Dependencies
mcp>=0.1.0

# Apple WeatherKit dependencies
requests>=2.31.0
authlib>=1.3.0
pytz>=2024.1
python-dotenv>=1.0.0

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
